requirements.txt
joblib
numpy
scikit-learn==1.6.1
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
mysql-connector-python
psycopg2-binary
gunicorn
python-decouple
flask-cors
